requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-common==1.1.28
azure-core==1.34.0
azure-identity==1.23.0
azure-search-documents==11.5.2
azure-storage-blob==12.25.1
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cors==1.0.1
cryptography==45.0.4
cssutils==2.11.1
distro==1.9.0
dnspython==1.16.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
future==1.0.0
gevent==25.5.1
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
more-itertools==10.7.0
msal==1.32.3
msal-extensions==1.3.1
openai==1.86.0
packaging==25.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==3.12.0
PySocks==1.7.1
python-dotenv==1.1.0
RapidFuzz==3.13.0
requests==2.32.4
requests-file==2.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
tldextract==5.3.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
crewai>=0.3.2
langchain>=0.2
pydantic>=2

